Add wordlists for Cosmos SDKo
Added wordlists for Cosmos SDK projects. In most cases the various
types of Msg structs are used as the scary strings. This list was
compiled from memory and from pasting some function names from the
Keepers in the documentation for Cosmos SDK. It hasn't been tested and
will likely need pruning and additions when worked with in practice.