add slopsquatting directory
Also add an entry for https://github.com/sashabaranov/go-openai.
This was suggested to me when I tried to connect Go to OpenAI.
The package might be fine but there should be a good reason to use it
over the official implementation: it shouldn't be the go-to for the LLM
to use.